pyproject.toml
[tool.poetry]
name = "focapd"
version = "0.1.0"
description = "Basic code to obtain probability distribution functions for EoL using TRI data"
authors = ["Jose D. Hernandez-Betancur <jodhernandezbemj@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]
exclude = ["ancillary", "data/raw"]


[project.urls]
"Homepage" = "https://github.com/jodhernandezbe/focapd-plastics"
"Bug Tracker" = "https://github.com/jodhernandezbe/focapd-plastics"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
typing = "^3.7.4.3"
pandas = "^2.1.4"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
sqlalchemy = "^2.0.36"
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"
pgmpy = "^0.1.26"
alembic = "^1.13.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
black = "^23.12.1"
flake8 = "^6.1.0"
isort = "^5.13.2"
interrogate = "^1.5.0"
pyright = "^1.1.369"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 128
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = true
ignore-nested-functions = false
ignore-nested-classes = true
ignore-setters = false
fail-under = 95
exclude = ["setup.py", "docs", "build"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 0
quiet = false
whitelist-regex = []
color = true

[tool.pyright]
include = ["src"]
extraPaths = ["src"] # for imports from other roots in subdirs
typeCheckingMode = "standard"
pythonVersion = "3.12"
reportMissingImports = true
reportMissingTypeStubs = false
stubPath = './typestubs'
venv = '.'
